// apps/web/__generated__/findUserByIdQuery.graphql.ts
export type Foot = "B" | "L" | "R" | "%future added value";
export type Gender = "M" | "W" | "%future added value";
export type Position = "CAM" | "CB" | "CDM" | "CF" | "CM" | "DF" | "FW" | "GK" | "LAM" | "LB" | "LCAM" | "LCB" | "LCM" | "LDM" | "LF" | "LM" | "LS" | "LW" | "LWB" | "MF" | "RAM" | "RB" | "RCAM" | "RCB" | "RCM" | "RDM" | "RF" | "RM" | "RS" | "RW" | "RWB" | "ST" | "SW" | "%future added value";
export type findUserByIdQuery$variables = {
  id: number;
};
export type findUserByIdQuery$data = {
  readonly findUserById: {
    readonly __typename: "UserModel";
    readonly activityArea: string | null | undefined;
    readonly birthDate: any | null | undefined;
    readonly email: string;
    readonly favoritePlayer: string | null | undefined;
    readonly foot: Foot | null | undefined;
    readonly gender: Gender | null | undefined;
    readonly id: number;
    readonly mainPosition: Position | null | undefined;
    readonly name: string | null | undefined;
    readonly phone: string | null | undefined;
    readonly preferredNumber: number | null | undefined;
    readonly profileImage: string | null | undefined;
    readonly provider: string | null | undefined;
    readonly subPositions: ReadonlyArray<Position> | null | undefined;
  };
};
export type findUserByIdQuery = {
  response: findUserByIdQuery$data;
  variables: findUserByIdQuery$variables;
};